fitur cart(belum selesai)#0.5.8

+++ database/migrations/2025_11_13_051824_create_order_items_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
             $table->id('order_item_id');
             $table->foreignId('order_id');
             $table->foreignId('item_id');
-            $table->integer('quantity');
-            $table->decimal('subtotal');
+            $table->smallInteger('quantity');
+            $table->decimal('subtotal', 10,2);
             $table->timestamps();
         });
     }
